renamed main file; added database access to main app, created dummy database, fixed things in the database model to make everything work as intented
